requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==4.0.1
Django==4.2.8
django-allauth==0.54.0
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.4
jiter==0.7.1
oauthlib==3.2.2
openai==1.54.4
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.4
